Move into adapter subdir
