+++ src/application.ts
@@ -3,9 +3,9 @@ import deprecated from 'deprecated-decorator'
 import express from 'express'
 import { EventEmitter } from 'promise-events'
 import { ApplicationFunction } from '.'
+import { GitHubApp } from './adapter/github-app'
 import { Context } from './context'
 import { GitHubAPI } from './github'
-import { GitHubApp } from './github-app'
 import { logger } from './logger'
 import { LoggerWithTarget, wrapLogger } from './wrap-logger'
